+++ resources/js/forms/form-element-types.ts
@@ -3,14 +3,12 @@ import CraftCombobox from '@craftcms/ui/vue/CraftCombobox.vue';
 import CraftEditableTable from '@craftcms/ui/vue/CraftEditableTable.vue';
 import CraftInput from '@craftcms/ui/vue/CraftInput.vue';
 import CraftInputMoney from '@craftcms/ui/vue/CraftInputMoney.vue';
-import CraftKeyedTable from '@craftcms/ui/vue/CraftKeyedTable.vue';
 import CraftObjectSelect from '@craftcms/ui/vue/CraftObjectSelect.vue';
 import CraftOptionRows from '@craftcms/ui/vue/CraftOptionRows.vue';
 import CraftSelect from '@craftcms/ui/vue/CraftSelect.vue';
 import CraftSwitch from '@craftcms/ui/vue/CraftSwitch.vue';
 import '@craftcms/ui/components/editable-table/editable-table';
 import '@craftcms/ui/components/input/input';
-import '@craftcms/ui/components/keyed-table/keyed-table';
 import '@craftcms/ui/components/object-select/object-select';
 import '@craftcms/ui/components/option-rows/option-rows';
 import '@craftcms/ui/components/select/select';
@@ -34,7 +32,6 @@ export const nativeFormElementRenderers = {
   'craft:editable-table-input': CraftEditableTable,
   'craft:element-condition-input': ElementConditionRenderer,
   'craft:field-layout-designer': FieldLayoutDesignerValueAdapter,
-  'craft:keyed-table-input': CraftKeyedTable,
   'craft:lightswitch-input': CraftSwitch,
   'craft:money-input': CraftInputMoney,
   'craft:number-input': CraftInput,